Introduce Ui class
